multimedia-review/requirements.txt
numpy==1.24.3

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 异步任务队列
celery==5.3.4
redis==5.0.1
flower==2.0.1

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0

# 文件处理
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
python-magic==0.4.27

# HTTP客户端
httpx==0.25.2
aiofiles==23.2.1

# 工具库
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
loguru==0.7.2

# 开发工具
pytest==7.4.3
